html2text.py: convert \n to <br> for the text in <blockquote>
Wordpress was rendering with <p> and <br> for the text in <blockquote>.
With this script, just convert \n to <br> instead of this.
The conversion should be skipped in the nested <pre> inside <blockquote>.

ex. /wp/archives/3499
 * exported text
```
<blockquote>^M
Hello,^M
^M
The current glue repository has been tagged as 1.0.11.^M
The highlights:^M
<ul>^M
<li>lrmd sets max number of children depending on the number of processors^M
</li>^M
<li>compatibility for stonith agents and hb_report for pacemaker v1.1.8^M
</li>^M
</ul>^M
</blockquote>^M
```

 * displayed HTML rendered by Wordpress
```
<blockquote><p>
Hello,</p>
<p>The current glue repository has been tagged as 1.0.11.<br />
The highlights:</p>
<ul>
<li>lrmd sets max number of children depending on the number of processors
</li>
<li>compatibility for stonith agents and hb_report for pacemaker v1.1.8
</li>
</ul>
</blockquote>
```